fix blob storage

+++ crop_disease_app/Models/Disease.cs
@@ -2,7 +2,7 @@ namespace hackathon_template.Models;
 
 public class Disease
 {
-    public Disease(int id, string userId, string diseaseName, decimal latitude, decimal longitude, string url) {
+    public Disease(int id, string userId, string diseaseName, double latitude, double longitude, string url) {
         Id = id;
         UserId = userId;
         DiseaseName = diseaseName;
@@ -14,7 +14,7 @@ public class Disease
     public int Id { get; set; }
     public string UserId { get; set; }
     public string DiseaseName { get; set; }
-    public decimal Latitude { get; set; }
-    public decimal Longitude { get; set; }
+    public double Latitude { get; set; }
+    public double Longitude { get; set; }
     public string Url { get; set; }
 }
